Fix twostream dummy variable issues. Fix LIDORT brdf counts to correct numbers.
Apparently in gfortran 4.8.5 a \ even inside of a comment will cause
the next line not to be parsed correctly. This was causing some
arguments to be declared as not dummy arguments. The intent was taken
out which broke other things. This change puts the intents back and
removes the \ causing the initial compiler errors.